// Timetic-organization-frontend/src/app/network/model/forAdminReport.ts
/**
 * Timetic szervezeti backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: you@your-company.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ForAdminReportAppointment } from './forAdminReportAppointment';


export interface ForAdminReport { 
    organizationName?: string;
    startDate?: number;
    endDate?: number;
    appointments?: Array<ForAdminReportAppointment>;
}